requirements.txt
Flask==2.3.2
line-bot-sdk==3.5.0
google-auth-oauthlib==1.0.0
google-auth==2.22.0
gspread==5.10.0
google-api-python-client==2.95.0
Pillow==10.0.0
gunicorn==21.2.0
